Added frontend login and register page to compliment the backend

+++ TaskManagerAPI/Controllers/TaskController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using TaskManagerAPI.Data;
 using TaskManagerAPI.Models;
+using System.Security.Claims;
 
 namespace TaskManagerAPI.Controllers
 {
@@ -22,7 +23,13 @@ namespace TaskManagerAPI.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<TaskItem>>> GetTasks()
         {
-            return await _context.TaskItems.ToListAsync();
+            int userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value!);
+            var userTasks = await _context.TaskItems
+                .Where(t => t.UserId == userId)
+                .ToListAsync();
+
+            //return await _context.TaskItems.ToListAsync();
+            return Ok(userTasks);
         }
 
         [HttpGet("{id}")]
@@ -32,14 +39,22 @@ namespace TaskManagerAPI.Controllers
             if (task == null) 
                 return NotFound();
 
+            int userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value!);
+            if (userId != task.UserId)
+                return Unauthorized();
+
             return task;
         }
 
         [HttpPost]
         public async Task<ActionResult<TaskItem>> CreateTask(TaskItem task)
         {
+            int userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value!);
+            task.UserId = userId;
+
             _context.TaskItems.Add(task);
             await _context.SaveChangesAsync();
+
             return CreatedAtAction(nameof(GetTask), new { id = task.Id }, task);
         }
 
@@ -47,8 +62,13 @@ namespace TaskManagerAPI.Controllers
         public async Task<IActionResult> UpdateTask(int id, TaskItem task)
         {
             if (id != task.Id) return BadRequest();
+
+            int userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value!);
+            task.UserId = userId;
+
             _context.Entry(task).State = EntityState.Modified;
             await _context.SaveChangesAsync();
+
             return NoContent();
         }
 
@@ -58,6 +78,10 @@ namespace TaskManagerAPI.Controllers
             var task = await _context.TaskItems.FindAsync(id);
             if (task == null) 
                 return NotFound();
+
+            int userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value!);
+            if (userId != task.UserId) 
+                return Unauthorized();
 
             _context.TaskItems.Remove(task);
             await _context.SaveChangesAsync();
